For #19988 - Part 1: Move RecentTabsListFeature into recenttabs package
